Fix scroll bug.
When the tilemap controls or the main map pane were scrolled, selection
and painting would choose/paint to the wrong places, because of the
scrollLeft and scrollTop data.

To ensure that these work correctly, scroll_state walks up the dom
from the node in question to the document root (whose parent node
is null) and sums all of their scrollTop and scrollLeft values.