Use canvases all over the darn place.
Rather than draw to images, and draw from images to canvases, and
hold the images around with big and small references, the code now
will just render the loaded images into at-scale canvases that are
held offscreen and when it needs to render one of those images to
the screen it will draw from those into new canvas elements that
can be scaled and can later be discarded.

This is a good first step of simplifying the image loading code before
adding in multi-layer drawing.