Add multilayer support.
Opted for a single canvas with multiple drawing passes -- this seems
like the less complicated way to go. Any number of layers can be
added and drawing is done on the current active layer (indicated by
the red highlight and the 'active_layer' class on the numbered layer-
select li's in the Tilemap Controls section, which are clickable to
change the layer that's being drawn to.

Saving/Loading fully supports the layers and will populate the layer
select buttons correctly, and support is maintained across source-
images when loading provided those images exist under the same names
in the current image_register.

Woo!