Kill all globals. (sort of)
Every variable used by every function should now be passed into that
function as an argument. The functions have largely been organised
into the order in which they are needed, and functions with limited
applicability have been pushed into the limited context in which they
apply.

The set of exposed symbols inside of 'main' is a bit silly and not set
according to any plan, and the 'setup_clicks' function is just a dumb
hodgepodge of all of the event listeners, because they have shared state
and have to mutate it and see its mutated results -- hiding them inside
of their own functions, e.g. setup_paint and setup_buttons for the two
different sets of responsibilities, would capture map at invocation of
those functions, and the listeners would bind on separate copies of the
map object. That's a paddlin'.

So they go together!

All in all, it's much better than it was -- and if I need to split it
into modules, it'll be far easier now.