Add play area.
Just renders the map and exposes the map load control -- actually
being able to play comes later.

I chose to duplicate the code, then remove all the unneccessary bits,
and regrettably make a few small changes, rather than try to refactor
the code to some non-duplicated state right off the bat. I'm not really
in this to build a glorious edifice of code quality: I just want to get
it working, so doing this lets my prioritize that. If the duplicated code
becomes a pain, I'll refactor it then.