Multi-canvas play drawing and keyboard control.
You can move around the marooned hero! (.. hahahaha)

My good pal Maroon Box has kindly agreed to stand in as The Hero until
we're able to get a circuit of regulars with things like faces and
transparent backgrounds hired.

I opted for three canvases: a base, a "player" canvas, and an overlay
canvas. The reason for this is to avoid having to figure out what squares
the hero is occluding, and subsequently iterating from the base layer to
the highest for all of them, every time the hero moves.

It's a very ugly implementation in some ways, because the layers that
constitute the 'base' and the 'overlay' are hard-coded into the looping
structures in the draw method :P -- I can imagine a way where I keep a
record of what layers belong to what context in the map, and then iterate
over the contexts and iterate over the layers within the contexts, but I
just didn't want to write it right now.

There's an interesting piece of complexity that I can foresee coming up:
sometimes the hero's head should draw over top of an impassable wall,
and sometimes the hero's legs should draw under the same impassable wall.
Say, if you had the top of a castle wall be something the hero could get
behind when walking around on the ground, their legs should draw under
the wall. But then they could walk around on top of the wall later, and
they should draw on top of it!

One way to solve this would be two different maps with things drawn at
different layer-heights. Another would be some complex logic about the
layer ordering vis-a-vis the different pieces of the hero. I'll think
some more about it.

Next up is making the hero purdy, then passability, and then... probably
some detours, but I think saving and teleporting are high on the priority
list.